fix: Fixed invalid channeltype and command renamed to "resolve".
